pyproject.toml
[tool.poetry]
name = "serafin"
version = "1.0.0"
description = "Logic-driven web content creation kit"
authors = ["SERAF <seraf-admin@medisin.uio.no>"]
license = "AGPL-3.0"

[tool.poetry.dependencies]
python = "^3.10"
django = ">=1.8,<3.0"
django-admin-sortable = "^2.3"
django-appconf = "^1.0.5"
django-compressor = "^4.0"
django-constance = "^2.9.0"
django-extensions = "^3.1.5"
django-filer = "^2.2.1"
django-import-export = "^2.7.1"
django-mptt = "^0.13.4"
django-request = "^1.6.1"
django-reversion = "^5.0.0"
django-settings-export = "^1.2.1"
django-sitetree = "^1.17.2"
django-storages = "^1.12.3"
django-suit = "^2.0a1"
django-user-agents = "^0.4.0"
django-xforwardedfor-middleware = "^2.0"
djangorestframework = "^3.13.1"
gevent = "^21.12.0"
gunicorn = "^20.1.0"
huey = "^2.4.3"
jsonfield = "^3.1.0"
mistune = "^2.0.2"
pillow = "^9.1.1"
psycopg2-binary = "2.8.6"
pyparsing = "^3.0.9"
redis = "^4.3.3"
requests = "^2.28.0"
rules = "^3.3"
weasyprint = "^55.0"
whitenoise = "^6.2.0"

[tool.poetry.dev-dependencies]
pylint = "^2.14.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
